requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
binance-connector==1.16.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
clikit==0.6.2
cmdstanpy==1.0.4
constantly==15.1.0
convertdate==2.4.0
crashtest==0.3.1
cryptography==37.0.4
cycler==0.11.0
Cython==0.29.32
ephem==4.1.3
fonttools==4.34.4
frozenlist==1.3.1
hijri-converter==2.2.4
holidays==0.14.2
httpstan==4.8.2
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
influxdb-client==1.31.0
kiwisolver==1.4.4
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
marshmallow==3.17.0
matplotlib==3.5.3
multidict==6.0.2
numpy==1.23.2
packaging==21.3
pandas==1.4.3
pastel==0.2.1
Pillow==9.2.0
prophet==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pylev==1.4.0
PyMeeus==0.5.11
pyOpenSSL==22.0.0
pyparsing==3.0.9
pysimdjson==3.2.0
pystan==2.19.1.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
Rx==3.2.0
service-identity==21.1.0
setuptools-git==1.2
six==1.16.0
tqdm==4.64.0
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.11
webargs==8.2.0
yarl==1.8.1
zope.interface==5.4.0
